+++ spec/dummy/app/models/barfoo.rb
@@ -1,3 +1,5 @@
 class Barfoo < ActiveRecord::Base
   include AbleToFailOnPurpose
+
+  validates :favorite_food, length: { maximum: 15 }
 end