+++ spec/dummy/app/permitters/special_barfoo_permitter.rb
@@ -1,0 +1,3 @@
+class SpecialBarfooPermitter < ActionController::Permitter
+  permit :id, :favorite_food
+end