+++ spec/dummy/app/serializers/minimal_barfoo_array_serializer.rb
@@ -1,0 +1,3 @@
+class MinimalBarfooArraySerializer < ActiveModel::ArraySerializer
+  self.root = false
+end