+++ spec/dummy/app/serializers/minimal_barfoo_serializer.rb
@@ -1,0 +1,3 @@
+class MinimalBarfooSerializer < ActiveModel::Serializer
+  attributes :favorite_food
+end