+++ spec/dummy/config/routes.rb
@@ -1,5 +1,5 @@
 Dummy::Application.routes.draw do
-  resources :barfoos do
+  resources :barfoos, :special_barfoos do
     get 'some_action', :on => :collection
   end